+++ src/test/groovy/org/hidetake/gradle/swagger/generator/ValidateSwaggerSpec.groovy
@@ -4,22 +4,6 @@ import spock.lang.Specification
 
 class ValidateSwaggerSpec extends Specification {
 
-    def "ValidateSwagger class should be available in a build script"() {
-        when:
-        def project = Fixture.projectWithPlugin()
-
-        then:
-        project.ValidateSwagger == ValidateSwagger
-    }
-
-    def "plugin should add default task"() {
-        when:
-        def project = Fixture.projectWithPlugin()
-
-        then:
-        project.tasks.findByName('validateSwagger')
-    }
-
     def "task should fail if YAML is wrong"() {
         given:
         def project = Fixture.projectWithPlugin {
